Some pi and controller functions (#13)
* Some pi and controller functions

* Fixed formatting issues in pidma.c and pimgr.c

* Changed devmgr.c label to match mdebug name

+++ src/io/pirawdma.c
@@ -3,7 +3,7 @@
 // TODO: this comes from a header
 #ident "$Revision: 1.17 $"
 
-s32 __osPiRawStartDma(s32 direction, u32 devAddr, void* dramAddr, size_t size) {
+s32 __osPiRawStartDma(s32 direction, u32 devAddr, void* dramAddr, u32 size) {
     register u32 stat;
 
     WAIT_ON_IOBUSY(stat);
